move get_score to puputils
